YAY-114 refactor: userEndpoints transfer from entities to shared, del usersEndpoints
