+++ src/entities/post/ui/Post/PostBody/PostDescription/PostDescription.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { Typography } from '@/shared/components/Typography'
 import { Avatar } from '@/shared/components/Avatar'
-import type { UserType } from '@/entities/user'
+import { UserType } from '@/entities/user/api'
 import { useRelativeTime } from '@/shared/hooks'
 import { ProfileLink } from '@/shared/components/ProfileLink'
