+++ src/entities/user/api/index.ts
@@ -1,3 +1,2 @@
-export * from './userEndpoints'
 export * from './user.service'
 export * from './user.types'
