--- src/entities/user/index.ts
@@ -1,4 +1,0 @@
-export { UserStats } from './ui/UserStats'
-export * from './api/user.types'
-export * from './api/user.service'
-export * from './api/userEndpoints'
