+++ src/entities/user/model/index.ts
@@ -1,1 +1,2 @@
 export * from './userProfileSchema'
+export * from './useUserProfile'
