--- src/entities/user/model/userEndpoints.ts
@@ -1,6 +1,0 @@
-export enum UserEndpoints {
-   usersTotalCount = '/public/users/total-count',
-   userProfile = '/public/users/profile',
-   settings = '/users/profile',
-   countriesWithCities = '/public/users/getCountriesWithCities',
-}
