+++ src/entities/user/ui/UserStats.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Typography } from '@/shared/components/Typography'
 
 type UserStatsProps = {
