+++ src/shared/enums/index.ts
@@ -1,5 +1,5 @@
 export * from './routes'
 export * from './authEndpoints'
 export * from './filesEndpoints'
-export * from './usersEndpoints'
+export * from './userEndpoints'
 export * from './publicPostsEndpoints'
