+++ src/shared/enums/userEndpoints.ts
