--- src/shared/enums/usersEndpoints.ts
@@ -1,3 +1,0 @@
-export enum UsersEndpoints {
-   profile = '/users/profile',
-}
